Export process-agent main function
