+++ cmd/process-agent/command/main_common.go
@@ -4,21 +4,17 @@
 // Copyright 2016-present Datadog, Inc.
 
 //nolint:revive // TODO(PROC) Fix revive linter
-package main
+package command
 
 import (
 	"context"
 	"fmt"
-	_ "net/http/pprof"
 	"os"
 
 	"go.uber.org/fx"
 
 	"github.com/DataDog/datadog-agent/cmd/agent/common/misconfig"
-	"github.com/DataDog/datadog-agent/cmd/internal/runcmd"
 	"github.com/DataDog/datadog-agent/cmd/manager"
-	"github.com/DataDog/datadog-agent/cmd/process-agent/command"
-	"github.com/DataDog/datadog-agent/cmd/process-agent/subcommands"
 	"github.com/DataDog/datadog-agent/comp/core"
 	"github.com/DataDog/datadog-agent/comp/core/config"
 	logComponent "github.com/DataDog/datadog-agent/comp/core/log"
@@ -46,7 +42,6 @@ import (
 	"github.com/DataDog/datadog-agent/pkg/tagger/local"
 	"github.com/DataDog/datadog-agent/pkg/tagger/remote"
 	ddutil "github.com/DataDog/datadog-agent/pkg/util"
-	"github.com/DataDog/datadog-agent/pkg/util/flavor"
 	"github.com/DataDog/datadog-agent/pkg/util/fxutil"
 	"github.com/DataDog/datadog-agent/pkg/util/log"
 	"github.com/DataDog/datadog-agent/pkg/version"
@@ -62,17 +57,7 @@ to your datadog.yaml file.
 Exiting.`
 )
 
-// main is the main application entry point
-func main() {
-	flavor.SetFlavor(flavor.ProcessAgent)
-
-	os.Args = command.FixDeprecatedFlags(os.Args, os.Stdout)
-
-	rootCmd := command.MakeCommand(subcommands.ProcessAgentSubcommands(), useWinParams, rootCmdRun)
-	os.Exit(runcmd.Run(rootCmd))
-}
-
-func runAgent(ctx context.Context, globalParams *command.GlobalParams) error {
+func runAgent(ctx context.Context, globalParams *GlobalParams) error {
 	if globalParams.PidFilePath != "" {
 		err := pidfile.WritePID(globalParams.PidFilePath)
 		if err != nil {
@@ -98,7 +83,7 @@ func runAgent(ctx context.Context, globalParams *command.GlobalParams) error {
 	return runApp(ctx, globalParams)
 }
 
-func runApp(ctx context.Context, globalParams *command.GlobalParams) error {
+func runApp(ctx context.Context, globalParams *GlobalParams) error {
 	exitSignal := make(chan struct{})
 	defer log.Flush() // Flush the log in case of an unclean shutdown
 	go util.HandleSignals(exitSignal)
@@ -121,7 +106,7 @@ func runApp(ctx context.Context, globalParams *command.GlobalParams) error {
 				),
 				ConfigParams: config.NewAgentParams(globalParams.ConfFilePath),
 				SecretParams: secrets.NewEnabledParams(),
-				LogParams:    command.DaemonLogParams,
+				LogParams:    DaemonLogParams,
 			},
 		),
 		// Populate dependencies required for initialization in this function
@@ -157,7 +142,7 @@ func runApp(ctx context.Context, globalParams *command.GlobalParams) error {
 		fxutil.FxLoggingOption(),
 
 		// Set `HOST_PROC` and `HOST_SYS` environment variables
-		fx.Invoke(command.SetHostMountEnv),
+		fx.Invoke(SetHostMountEnv),
 
 		// Initialize components not manged by fx
 		fx.Invoke(initMisc),
