+++ cmd/process-agent/command/main_windows.go
@@ -5,17 +5,16 @@
 
 //go:build windows
 
-package main
+package command
 
 import (
 	"context"
 	"errors"
 	"fmt"
-	_ "net/http/pprof"
+	_ "net/http/pprof" // Blank import used because this isn't directly used in this file
 	"os"
 	"time"
 
-	"github.com/DataDog/datadog-agent/cmd/process-agent/command"
 	"github.com/DataDog/datadog-agent/pkg/util/winutil"
 	"github.com/DataDog/datadog-agent/pkg/util/winutil/messagestrings"
 	"github.com/DataDog/datadog-agent/pkg/util/winutil/servicemain"
@@ -25,13 +24,14 @@ import (
 
 const (
 	// ServiceName is the service name used for the process-agent
-	ServiceName  = "datadog-process-agent"
-	useWinParams = true
+	ServiceName = "datadog-process-agent"
+	// UseWinParams is set to true when ran on Windows
+	UseWinParams = true
 )
 
 type service struct {
 	servicemain.DefaultSettings
-	globalParams *command.GlobalParams
+	globalParams *GlobalParams
 }
 
 func (s *service) Name() string {
@@ -56,7 +56,8 @@ func (s *service) Run(ctx context.Context) error {
 	return err
 }
 
-func rootCmdRun(globalParams *command.GlobalParams) {
+// RootCmdRun is the main function to run the process agent
+func RootCmdRun(globalParams *GlobalParams) {
 	var err error
 
 	if !globalParams.WinParams.Foreground {
