+++ cmd/process-agent/main.go
@@ -9,20 +9,21 @@
 package main
 
 import (
-	"context"
+	_ "net/http/pprof"
 	"os"
 
+	"github.com/DataDog/datadog-agent/cmd/internal/runcmd"
 	"github.com/DataDog/datadog-agent/cmd/process-agent/command"
+	"github.com/DataDog/datadog-agent/cmd/process-agent/subcommands"
+	"github.com/DataDog/datadog-agent/pkg/util/flavor"
 )
 
-const useWinParams = false
+// main is the main application entry point
+func main() {
+	flavor.SetFlavor(flavor.ProcessAgent)
 
-func rootCmdRun(globalParams *command.GlobalParams) {
-	// Invoke the Agent
-	err := runAgent(context.Background(), globalParams)
-	if err != nil {
-		// For compatibility with the previous cleanupAndExitHandler implementation, os.Exit() on error.
-		// This prevents runcmd.Run() from displaying the error.
-		os.Exit(1)
-	}
+	os.Args = command.FixDeprecatedFlags(os.Args, os.Stdout)
+
+	rootCmd := command.MakeCommand(subcommands.ProcessAgentSubcommands(), command.UseWinParams, command.RootCmdRun)
+	os.Exit(runcmd.Run(rootCmd))
 }
